(#17) Shared: merge the color-related code with the Avalonia branch
