--- src/SharedMath/Colors/AlphaChannelMode.cs
@@ -1,10 +1,0 @@
-namespace WpfMath.Colors
-{
-    /// <summary>Helps to differentiate RGBA and ARGB color models, if necessary.</summary>
-    internal enum AlphaChannelMode
-    {
-        None,
-        AlphaFirst,
-        AlphaLast
-    }
-}
