+++ src/SharedMath/SharedMath.projitems
@@ -9,51 +9,6 @@
     <Import_RootNamespace>SharedMath</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\AccentedAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\Atom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\BigOperatorAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\CharAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\CharSymbol.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\DummyAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\FencedAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\FixedCharAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\FractionAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\IRow.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\MatrixAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\NullAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\OverlinedAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\OverUnderDelimiter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\PhantomAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\Radical.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\RowAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\ScriptsAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\SpaceAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\StyledAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\SymbolAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\TypedAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\UnderlinedAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\UnderOverAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Atoms\VerticalCenteredAtom.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\Box.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\GlueBox.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\HorizontalBox.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\OverBar.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\OverUnderBox.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\StrutBox.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Boxes\VerticalBox.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\AlphaChannelMode.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\CmykColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\FixedComponentCountColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\FloatRgbColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\GrayscaleColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\HtmlColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\IColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\IntegerRgbColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\PredefinedColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\RgbaColor.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\RgbColorParserBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\SingleComponentColorParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Colors\StandardColorParsers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CharFont.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DefaultTexFont.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DelimiterFactory.cs" />
@@ -95,8 +50,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)TexFormulaHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TexPredefinedFormulaParser.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TexSymbolParser.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utils\ColorHelpers.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utils\Result.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)XmlUtilities.cs" />
   </ItemGroup>
 </Project>
