+++ src/WpfMath.Shared/Colors/AlphaChannelMode.cs
@@ -1,0 +1,9 @@
+namespace WpfMath.Colors;
+
+/// <summary>Helps to differentiate RGBA and ARGB color models, if necessary.</summary>
+internal enum AlphaChannelMode
+{
+    None,
+    AlphaFirst,
+    AlphaLast
+}
