+++ src/WpfMath.Shared/Colors/IColorParser.cs
@@ -1,13 +1,12 @@
 using System.Collections.Generic;
 
-namespace WpfMath.Colors
+namespace WpfMath.Colors;
+
+/// <summary>A parser for colors in commands like \color and \colorbox.</summary>
+public interface IColorParser
 {
-    /// <summary>A parser for colors in commands like \color and \colorbox.</summary>
-    public interface IColorParser
-    {
-        /// <summary>Parses the color components.</summary>
-        /// <param name="components">A sequence of the components that were separated by comma.</param>
-        /// <returns>Either a parsed color or <c>null</c> if it cannot be parsed.</returns>
-        RgbaColor? Parse(IEnumerable<string> components);
-    }
+    /// <summary>Parses the color components.</summary>
+    /// <param name="components">A list of the components that were separated by comma.</param>
+    /// <returns>Either a parsed color or <c>null</c> if it cannot be parsed.</returns>
+    RgbaColor? Parse(IReadOnlyList<string> components);
 }
