+++ src/WpfMath.Shared/Colors/StandardColorParsers.cs
@@ -1,20 +1,19 @@
 using System.Collections.Generic;
 
-namespace WpfMath.Colors
+namespace WpfMath.Colors;
+
+public static class StandardColorParsers
 {
-    public static class StandardColorParsers
+    public static IReadOnlyDictionary<string, IColorParser> Dictionary = new Dictionary<string, IColorParser>
     {
-        public static IReadOnlyDictionary<string, IColorParser> Dictionary = new Dictionary<string, IColorParser>
-        {
-            ["ARGB"] = new IntegerRgbColorParser(AlphaChannelMode.AlphaFirst),
-            ["argb"] = new FloatRgbColorParser(AlphaChannelMode.AlphaFirst),
-            ["cmyk"] = new CmykColorParser(),
-            ["gray"] = new GrayscaleColorParser(),
-            ["HTML"] = new HtmlColorParser(),
-            ["RGB"] = new IntegerRgbColorParser(AlphaChannelMode.None),
-            ["RGBA"] = new IntegerRgbColorParser(AlphaChannelMode.AlphaLast),
-            ["rgb"] = new FloatRgbColorParser(AlphaChannelMode.None),
-            ["rgba"] = new FloatRgbColorParser(AlphaChannelMode.AlphaLast)
-        };
-    }
+        ["ARGB"] = new IntegerRgbColorParser(AlphaChannelMode.AlphaFirst),
+        ["argb"] = new FloatRgbColorParser(AlphaChannelMode.AlphaFirst),
+        ["cmyk"] = new CmykColorParser(),
+        ["gray"] = new GrayscaleColorParser(),
+        ["HTML"] = new HtmlColorParser(),
+        ["RGB"] = new IntegerRgbColorParser(AlphaChannelMode.None),
+        ["RGBA"] = new IntegerRgbColorParser(AlphaChannelMode.AlphaLast),
+        ["rgb"] = new FloatRgbColorParser(AlphaChannelMode.None),
+        ["rgba"] = new FloatRgbColorParser(AlphaChannelMode.AlphaLast)
+    };
 }
